+++ src/app/components/planners/Planners.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { PlannersProvider } from "@contexts/PlannersProvider";
+import { PlannersProvider } from "@/app/contexts/PlannersProvider";
 
 import ExportModal from "../modals/exportModal/ExportModal";
 import PlannerList from "./PlannerList";
