--- src/app/contexts/MajorVerificationProvider/index.ts
@@ -1,4 +1,0 @@
-export {
-  MajorVerificationProvider,
-  MajorVerificationContext,
-} from "./Provider";
