+++ src/app/contexts/ModalsProvider.tsx
@@ -1,8 +1,26 @@
+import { Program } from "@/app/types/Program";
 import { PlannerContext } from "@contexts/PlannerProvider";
+import { SetShow, SetState } from "@customTypes/Common";
+import { PlannerData } from "@customTypes/Planner";
 import { createContext, useContext } from "react";
 
-import useModals from "../../hooks/useModals";
-import { ModalsContextProps } from "./Types";
+import useModals from "../hooks/useModals";
+
+export interface ModalsContextProps {
+  showExportModal: boolean;
+  setShowExportModal: SetShow;
+  showMajorsModal: boolean;
+  setShowMajorsModal: SetShow;
+  showPermissionsModal: boolean;
+  setShowPermissionsModal: SetShow;
+  showReplaceRLModal: boolean;
+  setShowReplaceRLModal: SetShow;
+  showMajorRequirementsEditModal: boolean;
+  setShowMajorRequirementsEditModal: SetShow;
+  courseState: PlannerData;
+  majorToEdit: Program | undefined;
+  setMajorToEdit: SetState<Program | undefined>;
+}
 
 export const ModalsContext = createContext({} as ModalsContextProps);
