--- src/app/contexts/ModalsProvider/index.ts
@@ -1,1 +1,0 @@
-export { ModalsProvider, ModalsContext } from "./Provider";
