+++ src/app/contexts/PermissionsProvider.tsx
@@ -1,7 +1,18 @@
+import { Permission } from "@/app/types/Permission";
+import { Program } from "@/app/types/Program";
 import { createContext } from "react";
 
-import usePermissions from "../../hooks/usePermissions";
-import { PermissionsContextProps } from "./Types";
+import usePermissions from "../hooks/usePermissions";
+
+export interface PermissionsContextProps {
+  loadingPermissions: boolean;
+  isSaved: boolean;
+  permissions: Permission[];
+  onUpsertPermission: (permission: Permission) => void;
+  onRemovePermission: (userEmail: string) => void;
+  isAdmin: boolean;
+  majorsAllowedToEdit: Program[];
+}
 
 export const PermissionsContext = createContext({} as PermissionsContextProps);
