--- src/app/contexts/PermissionsProvider/index.ts
@@ -1,1 +1,0 @@
-export { PermissionsProvider, PermissionsContext } from "./Provider";
