+++ src/app/contexts/PlannerProvider.tsx
@@ -1,10 +1,45 @@
+import { CustomCourseInput, StoredCourse } from "@customTypes/Course";
+import { Label } from "@customTypes/Label";
+import { PlannerData } from "@customTypes/Planner";
+import { DropResult } from "@hello-pangea/dnd";
 import { useSession } from "next-auth/react";
 import { createContext } from "react";
 
-import useCustomCourseSelection from "../../hooks/useCustomCourseSelection";
-import useHandleCourseDrag from "../../hooks/useHandleCourseDrag";
-import usePlanner from "../../hooks/usePlanner";
-import { PlannerContextProps, PlannerProviderProps } from "./Types";
+import useCustomCourseSelection from "../hooks/useCustomCourseSelection";
+import useHandleCourseDrag from "../hooks/useHandleCourseDrag";
+import usePlanner from "../hooks/usePlanner";
+
+export interface PlannerProviderProps {
+  children: React.ReactNode;
+  plannerId: string;
+  title: string;
+  order: number;
+}
+
+export interface PlannerContextProps {
+  deleteCourse: (quarterId: string) => (deleteIdx: number) => void;
+  editCustomCourse: (course: StoredCourse) => void;
+  totalCredits: number;
+  geSatisfied: string[];
+  courseState: PlannerData;
+  replaceCustomCourse: (customId: string, courses: StoredCourse[]) => void;
+  handleDragEnd: (result: DropResult) => void;
+  getCourseLabels: (course: StoredCourse) => Label[];
+  getAllLabels: () => Label[];
+  updatePlannerLabels: ({
+    labels,
+    newCourse,
+  }: {
+    labels: Label[];
+    newCourse?: StoredCourse;
+  }) => void;
+  customCourses: StoredCourse[];
+  handleAddCustom: (input: CustomCourseInput) => void;
+  handleRemoveCustom: (idx: number) => void;
+  updateNotes: (content: string) => void;
+  addYear: () => void;
+  deleteYear: (year: number) => void;
+}
 
 export const PlannerContext = createContext({} as PlannerContextProps);
