--- src/app/contexts/PlannerProvider/Types.ts
@@ -1,36 +1,0 @@
-import { CustomCourseInput, StoredCourse } from "@customTypes/Course";
-import { Label } from "@customTypes/Label";
-import { PlannerData } from "@customTypes/Planner";
-import { DropResult } from "@hello-pangea/dnd";
-
-export interface PlannerProviderProps {
-  children: React.ReactNode;
-  plannerId: string;
-  title: string;
-  order: number;
-}
-
-export interface PlannerContextProps {
-  deleteCourse: (quarterId: string) => (deleteIdx: number) => void;
-  editCustomCourse: (course: StoredCourse) => void;
-  totalCredits: number;
-  geSatisfied: string[];
-  courseState: PlannerData;
-  replaceCustomCourse: (customId: string, courses: StoredCourse[]) => void;
-  handleDragEnd: (result: DropResult) => void;
-  getCourseLabels: (course: StoredCourse) => Label[];
-  getAllLabels: () => Label[];
-  updatePlannerLabels: ({
-    labels,
-    newCourse,
-  }: {
-    labels: Label[];
-    newCourse?: StoredCourse;
-  }) => void;
-  customCourses: StoredCourse[];
-  handleAddCustom: (input: CustomCourseInput) => void;
-  handleRemoveCustom: (idx: number) => void;
-  updateNotes: (content: string) => void;
-  addYear: () => void;
-  deleteYear: (year: number) => void;
-}
