--- src/app/contexts/PlannerProvider/index.ts
@@ -1,2 +1,0 @@
-export { PlannerProvider, PlannerContext } from "./Provider";
-export { type PlannerProviderProps } from "./Types";
