+++ src/app/contexts/PlannersProvider.tsx
@@ -1,7 +1,32 @@
+import { PlannerData } from "@customTypes/Planner";
 import { createContext } from "react";
 
-import { usePlanners } from "../../hooks/usePlanners";
-import { PlannersContextProps, PlannersProviderProps } from "./Types";
+import { usePlanners } from "../hooks/usePlanners";
+import { SetState } from "../types/Common";
+
+export interface MultiPlanner {
+  planners: PlannerData[];
+  activePlanner: string | undefined;
+}
+
+export interface PlannersContextProps {
+  planners: PlannerData[];
+  removePlanner: (plannerId: string) => void;
+  addPlanner: () => void;
+  getPlanner: (id: string) => PlannerData;
+  setPlanner: (id: string, courseState: PlannerData) => void;
+  switchPlanners: (id: string) => void;
+  changePlannerName: (id: string, newTitle: string) => void;
+  replaceCurrentPlanner: () => void;
+  duplicatePlanner: (id: string) => void;
+  activePlanner: string | undefined;
+  showExportModal: boolean;
+  setShowExportModal: SetState<boolean>;
+}
+
+export interface PlannersProviderProps {
+  children: React.ReactNode;
+}
 
 export const PlannersContext = createContext({} as PlannersContextProps);
