--- src/app/contexts/PlannersProvider/Types.ts
@@ -1,27 +1,0 @@
-import { PlannerData } from "@customTypes/Planner";
-
-import { SetState } from "../../types/Common";
-
-export interface MultiPlanner {
-  planners: PlannerData[];
-  activePlanner: string | undefined;
-}
-
-export interface PlannersContextProps {
-  planners: PlannerData[];
-  removePlanner: (plannerId: string) => void;
-  addPlanner: () => void;
-  getPlanner: (id: string) => PlannerData;
-  setPlanner: (id: string, courseState: PlannerData) => void;
-  switchPlanners: (id: string) => void;
-  changePlannerName: (id: string, newTitle: string) => void;
-  replaceCurrentPlanner: () => void;
-  duplicatePlanner: (id: string) => void;
-  activePlanner: string | undefined;
-  showExportModal: boolean;
-  setShowExportModal: SetState<boolean>;
-}
-
-export interface PlannersProviderProps {
-  children: React.ReactNode;
-}
