--- src/app/contexts/PlannersProvider/index.ts
@@ -1,2 +1,0 @@
-export { PlannersProvider, PlannersContext } from "./Provider";
-export { type PlannersProviderProps, type MultiPlanner } from "./Types";
